app/requirements.txt
fastapi
uvicorn[standard]
neo4j
psycopg2-binary
pydantic
